+++ src/UI_Helper/RunningMode.h
@@ -14,7 +14,7 @@
 
 class RunningMode:public PropertyEdit {
 public:
-	RunningMode(LiquidCrystal& lcd_, int value);
+	RunningMode(LiquidCrystal& lcd_, float value);
 	virtual ~RunningMode();
 	void increment();
 	void decrement();
@@ -22,8 +22,8 @@ public:
 	void cancel();
 	void setFocus(bool focus);
 	void display();
-	void displayValue(int value);
-	void setValue(int value);
+	void displayValue(float value);
+	void setValue(float value);
 	std::string name();
 private:
 	std::string run ="Run now?";
@@ -31,8 +31,8 @@ private:
 	void save();
 	void displayEditValue();
 	LiquidCrystal& lcd;
-	int value;
-	int edit;
+	float value;
+	float edit;
 	bool focus;
 };
